chore(snapshot): swap CLI to tsx runner
